Decoupling from FastAPI with Client class, adding weekly median imputation...
