--- app/Dockerfile
@@ -1,17 +1,0 @@
-FROM python:3.12-slim
-
-WORKDIR /app
-
-# Install system dependencies
-RUN apt-get update && apt-get install -y git && rm -rf /var/lib/apt/lists/*
-
-# Install Python dependencies
-RUN pip install fastapi uvicorn \
-    && pip install https://github.com/dabadav/ai-cdss/releases/download/v0.1.1/ai_cdss-0.1.1.tar.gz
-
-# Copy API code
-COPY . .
-
-EXPOSE 8000
-
-CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000"]
