--- app/dependencies.py
@@ -1,5 +1,0 @@
-# dependencies.py
-from config import Settings
-
-def get_settings() -> Settings:
-    return Settings()
