+++ scripts/compute_ppf.py
@@ -35,8 +35,8 @@ Run this script from the command line:
 # %%
 from pathlib import Path
 import pandas as pd
-from ai_cdss.data_loader import DataLoaderLocal
-from ai_cdss.data_processor import ClinicalSubscales, ProtocolToClinicalMapper, compute_ppf, compute_protocol_similarity
+from ai_cdss.loaders import DataLoaderLocal
+from ai_cdss.processing import ClinicalSubscales, ProtocolToClinicalMapper, compute_ppf, compute_protocol_similarity
 
 def main():
