+++ src/ai_cdss/__init__.pyi
@@ -2,8 +2,8 @@
 
 # Public API for ai_cdss
 from .cdss import CDSS
-from .data_loader import DataLoader
-from .data_processor import DataProcessor, ClinicalSubscales, ProtocolToClinicalMapper
+from .loaders import DataLoader
+from .processing import DataProcessor, ClinicalSubscales, ProtocolToClinicalMapper
 
 __all__ = [
     "CDSS",
