+++ src/ai_cdss/constants.py
@@ -1,5 +1,6 @@
 # constants.py
 from pathlib import Path
+from enum import Enum
 
 ############################
 # DATA
@@ -13,6 +14,7 @@ DEFAULT_OUTPUT_DIR.mkdir(parents=True, exist_ok=True)
 DEFAULT_LOG_DIR = Path.home() / ".ai_cdss" / "logs"
 DEFAULT_LOG_DIR.mkdir(parents=True, exist_ok=True)
 
+DEFAULT_LOG_SCORING_FILEPATH = "scored_features_{scoring_date}.csv"
 PPF_PARQUET_FILEPATH = DEFAULT_OUTPUT_DIR / "ppf.parquet"
 
 ############################
@@ -24,7 +26,6 @@ SAVGOL_WINDOW_SIZE = 7
 SAVGOL_POLY_ORDER = 2
 THEILSON_REGRESSION_WINDOW_SIZE = 7
 
-
 ############################
 # ABBREVIATIONS
 ############################
@@ -32,6 +33,17 @@ PATIENT_ID = "PATIENT_ID"
 PROTOCOL_ID = "PROTOCOL_ID"
 PRESCRIPTION_ID = "PRESCRIPTION_ID"
 SESSION_ID = "SESSION_ID"
+CLINICAL_START = "CLINICAL_TRIAL_START_DATE"
+CLINICAL_END = "CLINICAL_TRIAL_END_DATE"
+
+STATUS = "STATUS"
+
+class SessionStatus(str, Enum):
+    CLOSED = "CLOSED"
+    NOT_PERFORMED = "NOT_PERFORMED"
+
+    def __str__(self):
+        return self.value
 
 ADHERENCE = "ADHERENCE"
 RECENT_ADHERENCE = "ADHERENCE_RECENT"
@@ -50,19 +62,37 @@ CONTRIB = "CONTRIB"
 USAGE = "USAGE"
 USAGE_WEEK = "USAGE_WEEK"
 DAYS = "DAYS"
+WEEKS_SINCE_START = "WEEKS_SINCE_START"
 
 SESSION_DATE = "SESSION_DATE"
+SESSION_INDEX = "SESSION_INDEX"
+SESSION_DURATION = "SESSION_DURATION"
+REAL_SESSION_DURATION = "REAL_SESSION_DURATION"
 GAME_MODE = "GAME_MODE"
 SECONDS_FROM_START = "SECONDS_FROM_START"
 WEEKDAY_INDEX = "WEEKDAY_INDEX"
+PRESCRIPTION_STARTING_DATE = "PRESCRIPTION_STARTING_DATE"
 PRESCRIPTION_ENDING_DATE = "PRESCRIPTION_ENDING_DATE"
 PRESCRIPTION_ACTIVE = "2100-01-01"
+PRESCRIBED_SESSION_DURATION = "PRESCRIBED_SESSION_DURATION"
+TOTAL_SUCCESS = "TOTAL_SUCCESS"
+TOTAL_ERRORS = "TOTAL_ERRORS"
+GAME_SCORE = "GAME_SCORE"
 
-SCORE = "GAME_SCORE"
+
+SCORE = "SCORE"
+
+# DM specific constants
+DM_SMOOTH = "DM_SMOOTH"
+
+# PPFSchema constants
+PROTOCOL_A = "PROTOCOL_A"
+PROTOCOL_B = "PROTOCOL_B"
+SIMILARITY = "SIMILARITY"
 
 # Session columns
 SESSION_COLUMNS: list[str] = [
-    "SESSION_ID", "TOTAL_SUCCESS", "TOTAL_ERRORS", "GAME_SCORE"
+    "SESSION_ID", "DM_VALUE", "TOTAL_SUCCESS", "TOTAL_ERRORS", "GAME_SCORE"
 ]
 METRIC_COLUMNS: list[str] = [
     "DM_VALUE", "PE_VALUE"
@@ -78,4 +108,4 @@ BY_PPST: list[str] = [PATIENT_ID, PROTOCOL_ID, SESSION_ID, SECONDS_FROM_START] #
 BY_ID: list[str] = BY_PPS + [PRESCRIPTION_ID] # By Patient–Protocol–Session + Prescription
 
 # Metrics
-FINAL_METRICS = [PPF, CONTRIB, RECENT_ADHERENCE, DELTA_DM, USAGE, USAGE_WEEK, DAYS, SCORE]
+FINAL_METRICS = [WEEKS_SINCE_START, PPF, CONTRIB, RECENT_ADHERENCE, DELTA_DM, USAGE, USAGE_WEEK, DAYS, SCORE]
