+++ src/ai_cdss/interface/__init__.py
@@ -1,0 +1,1 @@
+from .recommender import CDSSInterface
