+++ src/ai_cdss/processing/__init__.py
@@ -1,0 +1,8 @@
+from .processor import DataProcessor
+from .clinical import ClinicalSubscales, ProtocolToClinicalMapper
+
+__all__ = [
+    "DataProcessor",
+    "ClinicalSubscales",
+    "ProtocolToClinicalMapper",
+]
